+++ Python_examples/tecza_iteracyjnie.py
@@ -1,34 +1,53 @@
-# Kod źródłowy dla lekcji: Tęcza i dyskoteka - wersja z iteracją
+# Kod źródłowy dla lekcji:
+# Tęcza budowana z kolorów RGB - pętla for i pętla nieskończona - pełna gama kolorów
+
 # Autor: Adam Jurkiewicz, ABIX Edukacja https://pytechbrain.edu.pl
 #
+# template file to be used with PyTechBrain library
 from PyTechBrain import *
 from time import sleep
-from random import randint
-uklad = PyTechBrain()
+from sys import exit as exit_application
 
+# create object
+my_board = PyTechBrain()
 
+if my_board.board_init():  # initializing connection
+    print("Super! Initialization complete.")
+    my_board.set_buzzer("beep")  # demo, on, off
+else:
+    print("Something was wrong!")
+    my_board.full_debug_output()
+    my_board.usage_info()
+    exit_application(2)
+
+# we set off all outputs
+if not my_board.set_off_outputs():
+    print("Something was wrong!")
+    my_board.full_debug_output()
+    exit_application(2)
+
+# now we do the rest.... because everything works!
 while True:
-    print('Zapalam czerwoną')
+    print("RED on...")
     for liczba in range(255):
-        uklad.RGB_czerwona(liczba)
+        my_board.set_rgb_red(liczba)
         sleep(0.02)
 
-    print('Gaszę czerwoną')
-    uklad.RGB_czerwona(0)
+    print("RED off...")
+    my_board.set_rgb_red(0)
 
-    print('Zapalam niebieską')
+    print("GREEN on...")
     for liczba in range(255):
-        uklad.RGB_niebieska(liczba)
+        my_board.set_rgb_green(liczba)
         sleep(0.02)
 
-    print('Gaszę niebieską')
-    uklad.RGB_niebieska(0)
+    print("GREEN off...")
+    my_board.set_rgb_green(0)
 
-    print('Zapalam zieloną')
+    print("BLUE on...")
     for liczba in range(255):
-        uklad.RGB_zielona(liczba)
+        my_board.set_rgb_blue(liczba)
         sleep(0.02)
 
-    print('Gaszę zieloną')
-    uklad.RGB_zielona(0)
-
+    print("BLUE off...")
+    my_board.set_rgb_blue(0)