+++ Python_examples/template.py
@@ -3,4 +3,5 @@
 #
 from PyTechBrain import *
 from time import sleep
+
 uklad = PyTechBrain()