+++ PyTechBrain/changelog.md
@@ -4,4 +4,7 @@
 Initial version of new, redesigned library
 
 ## 0.8.2 - 8.08.2022:
-Added `full_reset(self) -> bool:` - clears all output to zero (black)
+Added `full_reset(self) -> bool` - clears all output to zero (black)
+
+## 0.8.3 - 20.08.2022:
+Added `list_devices_and_com_ports(self) -> json` - returns informations on all COM ports and available informations in JSON dumps